users_service-main/requirements.txt
Flask==2.3.2
SQLAlchemy==2.0.35
flask_caching
psycopg2
urllib3
google-auth
google-auth-oauthlib
oauthlib
requests
requests-oauthlib
python-jose[cryptography] 
gunicorn==22.0.0